pyproject.toml
[project]
name = "frameiq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.8.0",
    "attrs==25.1.0",
    "authlib==1.3.0",
    "chromadb>=0.5.5",
    "click==8.1.8",
    "cloudinary==1.41.0",
    "colorama==0.4.6",
    "email-validator==2.2.0",
    "flask==3.1.0",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.1",
    "groq==0.18.0",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "huggingface-hub==0.27.0",
    "langchain==0.3.14",
    "langchain-community==0.3.14",
    "langchain-core==0.3.30",
    "langchain-groq==0.2.3",
    "langchain-text-splitters==0.3.5",
    "langgraph==0.2.60",
    "langsmith==0.2.10",
    "numpy==1.26.4",
    "psycopg2-binary==2.9.9",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-oauthlib==1.3.1",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "sentence-transformers==3.3.1",
    "sqlalchemy==2.0.38",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "werkzeug==3.1.3",
    "wtforms==3.1.2",
]
